SFR-412 Improve link designations
This implements the boolean/flag link type checking on the front
end that has been implemented on the back end. This is a necessary
upgrade to prevent the React app from displaying "Download" for
all links as the `rel_type` field that previously controlled this
has been removed.

This does not implement further UI features such as the external link
icon, which should fall to SFR-413.