SFR-594 Add new analytics property
This adds the new ResearchNow-specific Google Analytics code, replacing the default NYPL code that is brought in through the `dgx-react-ga` component. That library is still used to load the code/provide other functionality. Two options for different ga codes are provided: `production` and `development` but one code is used for both for now. In the future it will be possible to specify different properties for different versions of the app