SFR-595 Add ga tracker for pages
As a single-page application ResearchNow requires a little additional configuration to track page views, and through that, search events. This can be done by hooking into the history object and listening to changes to the history, which can fire the ga tracker. With these events in ga the site search features should work automatically