SFR-595 Add search params to analytics tracker
This adds the URL parameters from `location.search` to the ga pageview tracker, which is necessary to track site searches. The URL parameters object is called `search` but seems to include all parameters, including for work pages. Unecessary parameters will be filtered out in ga, so this should not present a problem.