SFR-761 Fix non-keyword searches
The `onQueryChange` method was using the old, simple search-only, `query` style, which was leading to all queries being treated as keyword searches. This update ensures that the app looks in the right place for the query properly displays the "pills" for each search term, including searches that originated on the detail pages.