SFR-442 Add Feedback button
This adds a button for users to provide feedback on the ResearchNow project. This has been copied from the Shared Collections Catalog, it is the same component/styles, modified slightly to fit in with the ResearchNow app. It's use still needs to be cleared with that team.

This is an initial commit, it still needs to be wired to a backend system and re-styled to match the visual design of the new app