SFR-442 Add Radio Buttons and new styles
This adds the radio button question that is in the basic wireframe attached to the JIRA story. It also includes the necessary Privacy Policy link as we are asking for email addresses. This also adds some additional styling and a stub test class that just verifies that the component parses and loads properly.