SFR-593 Fix author name search
The method that constructs author searches form detail pages and search results was passing `name` as the query `field`. This should be `author` the method was returng the same field for both the `query` lookup and the `field` which works for VIAF and LCNAF queries but not for author name searches.

I've updated this to return a tuple, which is then accessed in the calling method. This fixes the issue and should "future-proof" this a bit in case we want to add future searches which may require different field and query names.

Also added are a few tests on the method to ensure that it returns the expected value.