HOTFIX Remove HTTP protocol coercion
This shim was put in to force URLs to use a certain protocol for environmental reasons (generally being forced to https). Due to updates in the environments this is no longer necessary.

The method is still in place but now simply adds a protocol to any URL that is missing one (this is necessary, for example, with covers). This should allow for certain components, such as the reader app, to work as intended in all environments.

This also adds a new environment: `local`. This is similar to `development` but has some specific settings to allow for easier local development. It can be started by using `npm run local-start` and requires that a local instance of the webpub-reader app be running at `localhost:4444`