SFR-798 Handle Single "Word" long strings
The string truncation method now handles very long strings that do not contain whitespace. These are simply truncated at the max length point. While this is very rare for titles, the main use case for this method now, it may be relevant in the future if this method is used to truncate shorter strings or other values since it is designed to be a general utility.