SFR-852 Add Citation Element
This adds a React component that takes a work and edition record and produces a formatted citation. Initially this is only the APA format, but other formats can be wrapped in the parser class to provide additional formats (MLA, Chicago, etc.)

This still has some work to be done, including:

- Comments/documentation
- Unit Testing
- Styling/Formatting

This is to provide a look at the basic implementation of this component and get feedback from the team.