SFR-852 Add Citations to Search Results and Work Details
This adds the new citation component to the Search Results and Work Detail pages behind a feature flag. They appear following every edition card. Right now the styling seems off and probably needs to be adjusted. Additionally these are not included on the Edition Detail page at the moment because that would require an adjustment to the API response for that page