SFR-852 Refactor Using feedback on organization
This is a refactor of the Citation component after receiving feedback on the initial commit. This organizes the logic of formatting each citation more modularly, with the data retrival methods being removed to a utility class.

This also simplifies the general system for both the component and the utility class.