HOTFIX AWS Elastic Beanstalk Environment Fix
A new/udpated npm package is causing a conflict with the AWS EBS environment. It is outputting unexpected non UTF-8 characters during the install process and this causes deployments to fail. This fix, which is supposedly the official recommendation from AWS is to pipe the output of the NPM install process to /dev/null. This removes the ability to easily review installation logs for troubleshooting, but does allow for deployments. Review should be made in the future to see if this can be removed.