NOREF Simplify and update EBS deployment process
This updates the repo to allow for deployment on EBS via next.js. This currently uses the dev server because the production server is not ready for deployment. This command uses the EBS `$PORT` variable to determine the port on which to serve the application (this is hidden in the EBS configuration so this is the recommended approach).

This also removes the `.ebextensions` folder since that contains custom configurations for the EBS environment that are no longer needed. An `.ebignore` has also been added to exclude some common files that aren't needed in the deployment.