NOREF Convert API URL to Environment Variable
The next.js application currently reads the API URL from a static configuration file committed to the repo. This makes temporary development/qa deployments difficult because the temporary API will generally be deployed to a non-standard URL (frequently a bare IP address).

This update updates the application to use a variable `API_URL` to configure this. It can be set in the `.env` file for local development or provided as a `-e API_URL=url` flag when running the the application in a container. This increases the portability of the containerized application and for a broader array of deployments