HOTFIX Update branch tags
This commit includes one fix and one improvement to the QA deployment process.

The fix is to correct a regression that was causing the `latest` tag for the application container to be deployed. The updated `task-definition` was not being passed in the correct step

The improvement is to tag images with the branch name instead of the git hash, this makes them more identifiable.