SFR-1073 Add Favicon to head
This simply sets the favicon link for the site in the `head` element. It does so alongside the `title` element which means that this should be easy to locate/update in the future if necessary.

The favicon URL is loaded from the `appConfig` file and should be changed there if the icon itself changes.