HOTFIX Container build tag
A recent change to how environment variables are handled required an update to how containers are built. As part of this process a YAML line was copied between files for GHA that contains a reference to an undefined variable. This resovles the issue by using a present value for the tag