HOTFIX Add production deployment step in production
The previous commit of the GHA file for production deployments is still stuck in development, which is an issue for current hotfix commits. This adds the file as a hotfix and should trigger builds on merging to development. This will likely cause a merge conflict that will have to be resolved in the near future.