SFR-1193 Add rspec regression tests
To do a better job of detecting regressions in the app this runs the rspec integration test suite against the current version deployed on QA. Since this will reflect the application version that will be deployed to production on the next release this should catch any/all regressions that affect usability.

Because this clones the rspec repo on each run it should stay in sync as that test suite continues to be improved.