SFR-1235 Load proxyUrl value on server side
The `proxyUrl` is provided to the new web reader component to enable the display of PDFs. This relies on server-side only environment variables to construct this value (specifically `API_URL`) and was unable to retrieve these values.

This updates the component by constructing the value on the server-side and passing it in as a new prop. A simple method constructs the string and should make it available in all cases.